+++ plone/app/customerize/tests/testDocTests.py
@@ -4,6 +4,16 @@ from plone.testing import layered
 from unittest import TestSuite
 
 import doctest
+import re
+import six
+
+
+class Py23DocChecker(doctest.OutputChecker):
+    def check_output(self, want, got, optionflags):
+        if six.PY2:
+            got = re.sub('zExceptions.NotFound', 'NotFound', got)
+            got = re.sub("u'(.*?)'", "'\\1'", want)
+        return doctest.OutputChecker.check_output(self, want, got, optionflags)
 
 
 def test_suite():
@@ -16,6 +26,7 @@ def test_suite():
                     testfile,
                     optionflags=OPTIONFLAGS,
                     package='plone.app.customerize.tests',
+                    checker=Py23DocChecker(),
                 ),
                 layer=PLONE_APP_CUSTOMERIZE_FUNCTIONAL_TESTING
             )